+++ config/yolo_2.ini
@@ -38,7 +38,7 @@ threshold = 0.1
 iou_threshold = 0.6
 anchors = [3.6577, 10.1917]
 class_names = ["tower"]
-checkpoint_path = ../out/yolo-50
+checkpoint_path = ../out/yolo-20
 cpu_only = True
 pretrained_weights_path = ../bin/yolov2.weights
