+++ net/yolo.py
@@ -1,248 +1,209 @@
+import abc
 import os
-import xml.etree.ElementTree as ET
 
-import cv2
-import imgaug as ia
 import numpy as np
-import sklearn.cluster
-from imgaug import augmenters as iaa
-from tqdm import tqdm
+import tensorflow as tf
 
-COLORS = [(0, 0, 255), (0, 255, 0), (255, 0, 0), (0, 255, 255), (255, 255, 0), (255, 0, 255)]
-
-# for img aug
-seq = iaa.Sequential([
-    iaa.Fliplr(0.5),
-    iaa.Flipud(0.5),
-    iaa.GaussianBlur((0, 3.0)),
-    iaa.Dropout(0.02),
-    iaa.AdditiveGaussianNoise(scale=0.01 * 255),
-    iaa.AdditiveGaussianNoise(loc=32, scale=0.0001 * 255),
-    iaa.Affine(translate_px={"x": (-40, 40)})
-])
+from . import base
+from . import v2
 
 
-def run_kmeans(data, num_anchors, tolerate, verbose=False):
-    km = sklearn.cluster.KMeans(n_clusters=num_anchors, tol=tolerate, verbose=verbose)
-    km.fit(data)
-    return km.cluster_centers_
+class Yolo(object):
+    @abc.abstractmethod
+    def create_network(self, num_anchors, num_classes, is_training, scope="yolo", input_shape=(416, 416, 3)):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def load_weights(self, layers, weights_path):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def find_bounding_boxes(self, out, anchors, threshold):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def generate_anchors(self, params):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def create_loss_fn(self, batch_size, net, anchors, class_names):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def make_batch(self, net, annotations, batch_size, anchors, class_names, augment_prob):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def create_train_optimizer(self, loss_fn, learning_rate):
+        raise NotImplementedError()
+
+    def test(self, params):
+        image_dir = params["image_dir"]
+        out_dir = params["out_dir"]
+        batch_size = int(params["batch_size"])
+        threshold = float(params["threshold"])
+        iou_threshold = float(params["iou_threshold"])
+        anchors = np.reshape(params["anchors"], [-1, 2])
+        class_names = params["class_names"]
+        input_h = int(params["input_h"])
+        input_w = int(params["input_w"])
+        input_c = int(params["input_c"])
+        checkpoint_path = params["checkpoint_path"]
+        pretrained_weights_path = params["pretrained_weights_path"]
+        cpu_only = params["cpu_only"].lower() == "true"
+
+        # load images
+        image_paths = base.load_image_paths(image_dir)
+        if len(image_paths) == 0:
+            print("No test images found in {}".format(image_dir))
+            return
+
+        # build network
+        net = self.create_network(len(anchors), len(class_names), False, input_shape=(input_h, input_w, input_c))
+
+        if cpu_only:
+            os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
+        with tf.Session() as sess:
+            sess.run(tf.global_variables_initializer())
+
+            # load checkpoint
+            saver = tf.train.Saver()
+            if not base.load_checkpoint_by_path(saver, sess, checkpoint_path):
+                # load pre-trained weights
+                ops = self.load_weights(net, pretrained_weights_path)
+                sess.run(ops)
+                print("Pre-trained weights loaded.")
+            else:
+                print("Checkpoint {} restored.".format(checkpoint_path))
+
+            input_shape = tuple(net[0].out.get_shape().as_list()[1:4])
+            test_batches = base.generate_test_batch(image_paths, batch_size, input_shape)
+            for x_batch, paths in test_batches:  # run batch
+                net_out = sess.run(net[-1].out, feed_dict={net[0].out: x_batch})
+
+                # post-process
+                net_boxes = []
+                net_out = np.reshape(net_out,
+                                     [-1, net_out.shape[1], net_out.shape[2], len(anchors), (5 + len(class_names))])
+                for out in net_out:
+                    bounding_boxes = self.find_bounding_boxes(out, anchors, threshold)
+                    net_boxes.append(base.non_maximum_suppression(bounding_boxes, iou_threshold))
+
+                for boxes, path in zip(net_boxes, paths):
+                    # draw box on image
+                    new_img = base.draw_boxes(path, boxes, class_names)
+                    # write to file
+                    file_name, file_ext = os.path.splitext(os.path.basename(path))
+                    out_path = os.path.join(out_dir, "{}_out{}".format(file_name, file_ext))
+                    base.save_image(new_img, out_path)
+                    print("{}: Found {} objects. Saved to {}".format(file_name, len(boxes), out_path))
+            print("Done")
+
+    def train(self, params):
+        train_image_dir = params["image_dir"]
+        train_annotation_dir = params["annotation_dir"]
+        val_image_dir = params["val_image_dir"]
+        val_annotation_dir = params["val_annotation_dir"]
+        batch_size = int(params["batch_size"])
+        learning_rate = float(params["learning_rate"])
+        augment_prob = float(params["augment_probability"])
+        checkpoint_prefix = params["checkpoint_prefix"]
+        checkpoint_dir = params["checkpoint_dir"]
+        checkpoint_step = int(params["checkpoint_step"])
+        pretrained_weights_path = params["pretrained_weights_path"]
+        tensorboard_log_dir = params["tensorboard_log_dir"]
+        anchors = np.reshape(params["anchors"], [-1, 2])
+        class_names = params["class_names"]
+        input_h = int(params["input_h"])
+        input_w = int(params["input_w"])
+        input_c = int(params["input_c"])
+        epochs = int(params["epochs"])
+        max_step = int(params["max_step"])
+        cpu_only = params["cpu_only"].lower() == "true"
+
+        # prepare data
+        train_annotations = base.parse_annotations(train_annotation_dir, train_image_dir)
+        assert len(train_annotations) > 0
+        val_annotations = base.parse_annotations(val_annotation_dir, val_image_dir)
+
+        # build network
+        net = self.create_network(len(anchors), len(class_names), True, input_shape=(input_h, input_w, input_c))
+
+        if cpu_only:
+            os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
+        with tf.Session() as sess:
+            sess.run(tf.global_variables_initializer())
+
+            # create loss function
+            loss, placeholders = self.create_loss_fn(batch_size, net, anchors, class_names)
+            train_op = self.create_train_optimizer(loss, learning_rate)
+            saver = tf.train.Saver()
+            with tf.Session() as sess:
+                tf_summary = tf.summary.merge_all()
+                train_writer = tf.summary.FileWriter(os.path.join(tensorboard_log_dir, "train"), sess.graph)
+                val_writer = tf.summary.FileWriter(os.path.join(tensorboard_log_dir, "validation"), sess.graph)
+                sess.run(tf.global_variables_initializer())
+
+                # load pre-trained weights/checkpoint
+                step = base.load_checkpoint(saver, sess, checkpoint_dir, checkpoint_prefix)
+                if step < 0:
+                    # load pre-trained weights
+                    ops = self.load_weights(net, pretrained_weights_path)
+                    sess.run(ops)
+                    print("Pre-trained weights loaded.")
+                    step = 0
+                    base.save_checkpoint(saver, sess, checkpoint_dir, checkpoint_prefix, step)
+                else:
+                    print("Checkpoint restored. step:{}".format(step))
+
+                # make train batch
+                train_loss_mva = None
+                for epoch in range(1, epochs + 1):
+                    if 0 <= max_step < step:
+                        break
+                    batches = self.make_batch(net, train_annotations, batch_size, anchors, class_names, augment_prob)
+                    for feed_images, feed_gts in batches:
+                        step += 1
+                        if 0 <= max_step < step:
+                            break
+                        feed_dict = {placeholders[key]: feed_gts[key] for key in placeholders}
+                        feed_dict[net[0].out] = feed_images
+                        _, summary, train_loss = sess.run([train_op, tf_summary, loss], feed_dict=feed_dict)
+
+                        train_writer.add_summary(summary, step)
+                        train_writer.flush()
+                        train_loss_mva = train_loss_mva * 0.9 + train_loss * 0.1 if train_loss_mva is not None else train_loss
+                        print("step {} ({}/{}): {} (moving average: {})".format(step, epoch, epochs, train_loss,
+                                                                                train_loss_mva))
+                        if step > 0 and step % checkpoint_step == 0:
+                            base.save_checkpoint(saver, sess, checkpoint_dir, checkpoint_prefix, step)
+
+                            # validation
+                            val_batches = self.make_batch(net, val_annotations, batch_size, anchors, class_names, 0)
+                            val_total = 0
+                            val_count = 0
+                            for val_feed_images, val_feed_gts in val_batches:
+                                val_feed_dict = {placeholders[key]: val_feed_gts[key] for key in placeholders}
+                                val_feed_dict[net[0].out] = val_feed_images
+                                val_loss = sess.run(loss, feed_dict=val_feed_dict)
+                                val_total += float(val_loss)
+                                val_count += 1
+                            val_loss = val_total / val_count
+                            val_summary = tf.Summary(value=[
+                                tf.Summary.Value(tag="loss", simple_value=val_loss),
+                            ])
+                            val_writer.add_summary(val_summary)
+                            val_writer.flush()
+                            print("validation loss: {}".format(val_loss))
+                    print("Epoch ({}/{}) completed.".format(epoch, epochs))
+        print("Done")
 
 
-def load_checkpoint_by_path(saver, sess, checkpoint_path):
-    try:
-        saver.restore(sess, checkpoint_path)
-        return True
-    except ValueError as e:
-        print("Failed to load {}: {}".format(checkpoint_path, str(e)))
-        return False
-
-
-def load_image_paths(path_to_img_dir):
-    return [os.path.join(os.path.abspath(path_to_img_dir), f) for f in os.listdir(path_to_img_dir)
-            if any(f.lower().endswith(ext) for ext in ["jpg", "bmp", "png", "gif"])]
-
-
-def parse_annotations(annotation_dir, image_dir, normalize=False):
-    annotations = [os.path.join(os.path.abspath(annotation_dir), f) for f in os.listdir(annotation_dir)
-                   if f.lower().endswith(".xml")]
-
-    result = []
-    for annotation in tqdm(annotations):
-        root = ET.parse(annotation).getroot()
-        img_path = os.path.join(image_dir, root.find("filename").text)
-
-        size = root.find("size")
-        w = int(size.find("width").text)
-        h = int(size.find("height").text)
-
-        img_objects = []
-        objects = root.findall("object")
-        for object in objects:
-            name = object.find("name").text
-            bndbox = object.find("bndbox")
-            x1 = int(bndbox.find("xmin").text)
-            y1 = int(bndbox.find("ymin").text)
-            x2 = int(bndbox.find("xmax").text)
-            y2 = int(bndbox.find("ymax").text)
-            if normalize:
-                x1, x2 = x1 / w, x2 / w
-                y1, y2 = y1 / h, y2 / h
-            img_objects.append((x1, y1, x2, y2, name))
-
-        result.append((img_path, img_objects))
-    return result
-
-
-def augment_image(image, objects, new_shape):
-    seq_det = seq.to_deterministic()
-
-    bb_objects = [ia.BoundingBox(x1=o[0], y1=o[1], x2=o[2], y2=o[3], label=o[4]) for o in objects]
-    bbs = ia.BoundingBoxesOnImage(bb_objects, shape=new_shape)
-
-    image_aug = seq_det.augment_images([image])[0]
-    bbs_aug = seq_det.augment_bounding_boxes([bbs])[0]
-
-    new_objects = [(o.x1, o.y1, o.x2, o.y2, o.label) for o in
-                   # bbs_aug.bounding_boxes]
-                   bbs_aug.remove_out_of_image().cut_out_of_image().bounding_boxes]
-    return image_aug, new_objects
-
-
-def preprocess_image(image_path, new_shape, objects=None, augment_prob=0.):
-    image = cv2.imread(image_path)
-    if image is None:
-        print("Failed to read {}".format(image_path))
-        return
-
-    net_image = cv2.resize(image, new_shape[0:2])
-    net_image = net_image[:, :, ::-1]  # BGR to RGB
-
-    new_objects = None
-    if objects is not None:
-        new_objects = []
-        for obj in objects:
-            xmin = obj[0] / image.shape[1] * new_shape[1]
-            ymin = obj[1] / image.shape[0] * new_shape[0]
-            xmax = obj[2] / image.shape[1] * new_shape[1]
-            ymax = obj[3] / image.shape[0] * new_shape[0]
-            new_objects.append((xmin, ymin, xmax, ymax, obj[4]))
-
-        if np.random.uniform() < augment_prob:
-            net_image, new_objects = augment_image(net_image, new_objects, new_shape)
-
-    # For debug only
-    """
-    from copy import deepcopy
-    tmp = deepcopy(net_image)
-    for obj in new_objects:
-        tl = int(obj[0]), int(obj[1])
-        br = int(obj[2]), int(obj[3])
-        tmp = np.ascontiguousarray(tmp[:, :, ::-1], dtype=np.uint8)
-        print(tl, br, tmp.shape)
-        print(tmp)
-        cv2.rectangle(tmp, tl, br, (0, 255, 0), thickness=3)
-        print(tl, br, tmp.shape)
-    cv2.imshow("image", tmp)
-    cv2.waitKey(0)
-    """
-
-    net_image = net_image / 255.
-
-    return net_image, new_objects
-
-
-def generate_test_batch(img_paths, batch_size, input_shape):
-    total_batches = int(np.ceil(len(img_paths) / batch_size))
-
-    for b in range(total_batches):
-        images = []
-        paths = []
-        for i in range(min(batch_size, len(img_paths) - b * batch_size)):
-            image, _ = preprocess_image(img_paths[b * batch_size + i], input_shape, augment_prob=0.)
-            images.append(np.expand_dims(image, axis=0))
-            paths.append(img_paths[b * batch_size + i])
-        yield np.concatenate(images, axis=0), paths
-
-
-def sigmoid(x):
-    return 1. / (1. + np.exp(-x))
-
-
-def softmax(x):
-    e_x = np.exp(x - np.max(x))
-    return e_x / e_x.sum()
-
-
-def iou_score(box1, box2):
-    box1_min, box1_max = box1.get_top_left(), box1.get_bottom_right()
-    box1_area = box1.w * box1.h
-    box2_min, box2_max = box2.get_top_left(), box2.get_bottom_right()
-    box2_area = box2.w * box2.h
-
-    intersect_min = np.maximum(box1_min, box2_min)
-    intersect_max = np.minimum(box1_max, box2_max)
-    intersect_wh = np.maximum(intersect_max - intersect_min, 0)
-    intersect_area = intersect_wh[0] * intersect_wh[1]
-    union_area = np.maximum(box1_area + box2_area - intersect_area, 1e-8)
-
-    return intersect_area / union_area
-
-
-def non_maximum_suppression(boxes, iou_threshold):
-    if len(boxes) == 0:
-        return []
-
-    boxes.sort(key=lambda box: box.prob, reverse=True)  # sort by prob(confidence)
-    new_boxes = [boxes[0]]  # add first element
-    for i in range(1, len(boxes)):
-        overlapping = False
-        for new_box in new_boxes:
-            if iou_score(new_box, boxes[i]) >= iou_threshold:
-                overlapping = True
-                break
-        if not overlapping:
-            new_boxes.append(boxes[i])
-    return new_boxes
-
-
-def draw_boxes(path_to_img, boxes, class_names):
-    image = cv2.imread(path_to_img)
-    assert image is not None
-    h, w = image.shape[0:2]
-    for box in boxes:
-        tl = np.maximum(box.get_top_left(h, w), 0)
-        br = np.maximum(box.get_bottom_right(h, w), 0)
-        tl = (int(tl[0]), int(tl[1]))
-        br = (int(br[0]), int(br[1]))
-
-        cv2.rectangle(image, tl, br, COLORS[box.class_idx % len(COLORS)], thickness=3)
-        class_name = class_names[box.class_idx]
-        cv2.putText(image, "{} {:.3f}".format(class_name, box.prob), (tl[0], tl[1] - 10), \
-                    cv2.FONT_HERSHEY_SIMPLEX, 0.5, COLORS[box.class_idx % len(COLORS)], thickness=1)
-    return image
-
-
-def save_image(image, out_path):
-    cv2.imwrite(out_path, image)
-
-
-def load_checkpoint(saver, sess, checkpoint_dir, checkpoint_prefix):
-    metas = [f for f in os.listdir(checkpoint_dir) if f.startswith(checkpoint_prefix) and f.endswith(".meta")]
-
-    if len(metas) == 0:
-        return -1
-
-    try:
-        metas.sort(reverse=True)
-        checkpoint_path = os.path.join(checkpoint_dir, os.path.splitext(metas[0])[0])
-        name = os.path.splitext(os.path.basename(checkpoint_path))[0]
-        step = int(name.split("-")[1])
-        saver.restore(sess, checkpoint_path)
-        return step
-    except ValueError as e:
-        print("Failed to load checkpoint: {}".format(str(e)))
-        return -1
-
-
-def save_checkpoint(saver, sess, checkpoint_dir, checkpoint_prefix, step):
-    out_path = os.path.join(checkpoint_dir, "{}-{}".format(checkpoint_prefix, step))
-    saver.save(sess, out_path)
-    print("Checkpoint saved to {}".format(out_path))
-
-
-class BoundingBox(object):
-    def __init__(self, x=0., y=0., w=0., h=0., cx=0, cy=0, class_idx=-1, prob=-1.):
-        self.x = x
-        self.y = y
-        self.w = w
-        self.h = h
-        self.cx = cx
-        self.cy = cy
-        self.class_idx = class_idx
-        self.prob = prob
-
-    def get_top_left(self, h=1., w=1.):
-        return (self.x - self.w / 2.) * w, (self.y - self.h / 2.) * h
-
-    def get_bottom_right(self, h=1., w=1.):
-        return (self.x + self.w / 2.) * w, (self.y + self.h / 2.) * h
+class YoloV2(Yolo):
+    create_network = v2.create_full_network
+    load_weights = v2.load_weights
+    find_bounding_boxes = v2.find_bounding_boxes
+    generate_anchors = v2.generate_anchors
+    make_batch = v2.make_batch
+    create_train_optimizer = v2.create_train_optimizer
+    create_loss_fn = v2.create_loss_fn
